Rename functions project and split apart
