+++ src/Microsoft.Health.Dicom.Core/Properties/AssemblyInfo.cs
@@ -8,10 +8,10 @@ using System.Resources;
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Microsoft.Health.Dicom.Core.UnitTests")]
-[assembly: InternalsVisibleTo("Microsoft.Health.Dicom.Functions")]
 [assembly: InternalsVisibleTo("Microsoft.Health.Dicom.Functions.Client")]
 [assembly: InternalsVisibleTo("Microsoft.Health.Dicom.Functions.Client.UnitTests")]
 [assembly: InternalsVisibleTo("Microsoft.Health.Dicom.Functions.UnitTests")]
+[assembly: InternalsVisibleTo("Microsoft.Health.Dicom.Operations")]
 [assembly: InternalsVisibleTo("Microsoft.Health.Dicom.SqlServer")]
 [assembly: InternalsVisibleTo("Microsoft.Health.Dicom.SqlServer.UnitTests")]
 [assembly: InternalsVisibleTo("Microsoft.Health.Dicom.Web.Tests.E2E")]
