+++ src/Microsoft.Health.Dicom.Operations/Configuration/DicomFunctionsConfiguration.cs
@@ -3,12 +3,10 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-namespace Microsoft.Health.Dicom.Functions.Configuration
+namespace Microsoft.Health.Dicom.Operations.Configuration
 {
     internal static class DicomFunctionsConfiguration
     {
-        public const string HostSectionName = "AzureFunctionsJobHost";
-
         public const string SectionName = "DicomFunctions";
     }
 }
