+++ src/Microsoft.Health.Dicom.Operations/Extensions/DurableFunctionsContextExtensions.cs
@@ -7,7 +7,7 @@ using System;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Microsoft.Health.Dicom.Core.Models.Operations;
 
-namespace Microsoft.Health.Dicom.Functions.Extensions
+namespace Microsoft.Health.Dicom.Operations.Extensions
 {
     internal static class DurableFunctionsContextExtensions
     {
