+++ src/Microsoft.Health.Dicom.Operations/Indexing/ReindexDurableFunction.Orchestration.cs
@@ -12,10 +12,10 @@ using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using Microsoft.Extensions.Logging;
 using Microsoft.Health.Dicom.Core.Features.ExtendedQueryTag;
 using Microsoft.Health.Dicom.Core.Features.Model;
-using Microsoft.Health.Dicom.Functions.Extensions;
-using Microsoft.Health.Dicom.Functions.Indexing.Models;
+using Microsoft.Health.Dicom.Operations.Extensions;
+using Microsoft.Health.Dicom.Operations.Indexing.Models;
 
-namespace Microsoft.Health.Dicom.Functions.Indexing
+namespace Microsoft.Health.Dicom.Operations.Indexing
 {
     public partial class ReindexDurableFunction
     {
